+++ src/BDSCOD/Types.hs
@@ -33,10 +33,14 @@ module BDSCOD.Types
   , PDESolution(..)
   , LogLikelihood
   , LlhdAndNB
-  , LlhdCalcState) where
+  , LlhdCalcState
+  , MCMCConfiguration(..)
+  , MWCSeed
+  ) where
 
 import Control.DeepSeq
 import Data.Aeson
+import qualified Data.Aeson as Json
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Builder as BBuilder
 import qualified Data.ByteString.Lazy as BL
@@ -44,10 +48,31 @@ import qualified Data.Csv as Csv
 import Data.List (intersperse)
 import Epidemic.Types.Parameter
 import GHC.Generics (Generic)
+import GHC.Word (Word32(..))
 
--- | TODO Give this orphan a home in @epi-sim@.
+-- | TODO Remove this once the version of @epi-sim@ has been updated to the
+-- latest one on github.
 instance Ord TimeDelta where
   (TimeDelta a) <= (TimeDelta b) = a <= b
+
+-- | Alias for the type used to seed the MWC PRNG.
+type MWCSeed = Word32
+
+-- | These objects configure an MCMC run.
+data MCMCConfiguration =
+  MCMCConfiguration
+    { -- | Where to write MCMC samples
+      mcmcOutputCSV :: FilePath
+      -- | The number of samples to generate
+    , mcmcNumIters :: Int
+      -- | The standard deviation of the step size
+    , mcmcStepSD :: Double
+      -- | The seed for the PRNG
+    , mcmcSeed :: MWCSeed
+    }
+  deriving (Show, Generic)
+
+instance Json.FromJSON MCMCConfiguration
 
 -- | The parameters of the constant rate BDSCOD are the birth rate, the natural
 -- removal rate, the sampling rate, the timing and probability of catastrophic
@@ -107,7 +132,12 @@ getNus :: Parameters -> Timed Probability
 getNus (Parameters (_, _, _, _, _, tns)) = tns
 
 type UnpackedParameters
-   = (Rate, Rate, Rate, [(AbsoluteTime, Probability)], Rate, [(AbsoluteTime, Probability)])
+   = ( Rate
+     , Rate
+     , Rate
+     , [(AbsoluteTime, Probability)]
+     , Rate
+     , [(AbsoluteTime, Probability)])
 
 unpackParameters :: Parameters -> UnpackedParameters
 unpackParameters (Parameters (pLambda, pMu, pPsi, Timed pRhos, pOmega, Timed pNus)) =
@@ -117,12 +147,11 @@ packParameters :: UnpackedParameters -> Parameters
 packParameters (pLambda, pMu, pPsi, pRhos, pOmega, pNus) =
   Parameters (pLambda, pMu, pPsi, Timed pRhos, pOmega, Timed pNus)
 
-
 -- | Return the times of scheduled events: catastrophes and disasters.
-scheduledTimes :: Parameters -> ([AbsoluteTime],[AbsoluteTime])
+scheduledTimes :: Parameters -> ([AbsoluteTime], [AbsoluteTime])
 scheduledTimes (Parameters (_, _, _, Timed pRhos, _, Timed pNus)) =
   let times = map fst
-  in (times pRhos, times pNus)
+   in (times pRhos, times pNus)
 
 -- | The number of lineages that exist in a phylogeny
 type NumLineages = Double
@@ -178,34 +207,36 @@ updateDelay (_, oEvent) delay = (delay, oEvent)
 
 -- | Predicate for the observation referring to a birth.
 isBirth :: Observation -> Bool
-isBirth = (==OBirth) . snd
+isBirth = (== OBirth) . snd
 
 -- | Predicate for the observation referring to an unscheduled and sequenced
 -- observation.
 isUnscheduledSequenced :: Observation -> Bool
-isUnscheduledSequenced = (==ObsUnscheduledSequenced) . snd
+isUnscheduledSequenced = (== ObsUnscheduledSequenced) . snd
 
 -- | Predicate for the observation referring to an occurrence.
 isOccurrence :: Observation -> Bool
-isOccurrence = (==OOccurrence) . snd
+isOccurrence = (== OOccurrence) . snd
 
 -- | The number of /unsequenced/ lineages that were observed.
 numUnsequenced :: Observation -> NumLineages
-numUnsequenced obs = case snd obs of
-  OBirth -> 0
-  ObsUnscheduledSequenced -> 0
-  OOccurrence -> 1
-  (OCatastrophe _) -> 0
-  (ODisaster n) -> n
+numUnsequenced obs =
+  case snd obs of
+    OBirth -> 0
+    ObsUnscheduledSequenced -> 0
+    OOccurrence -> 1
+    (OCatastrophe _) -> 0
+    (ODisaster n) -> n
 
 -- | The number of /sequenced/ lineages that were observed.
 numSequenced :: Observation -> NumLineages
-numSequenced obs = case snd obs of
-  OBirth -> 0
-  ObsUnscheduledSequenced -> 1
-  OOccurrence -> 0
-  (OCatastrophe n) -> n
-  (ODisaster _) -> 0
+numSequenced obs =
+  case snd obs of
+    OBirth -> 0
+    ObsUnscheduledSequenced -> 1
+    OOccurrence -> 0
+    (OCatastrophe n) -> n
+    (ODisaster _) -> 0
 
 -- | The negative binomial distribution extended to include the limiting case of
 -- a point mass at zero. The parameterisation is in terms of a positive
@@ -231,12 +262,11 @@ instance Csv.ToField NegativeBinomial where
 
 instance Csv.ToRecord NegativeBinomial
 
-data PDESolution = PDESol NegativeBinomial NumLineages
+data PDESolution =
+  PDESol NegativeBinomial NumLineages
 
 type LogLikelihood = Double
 
-type LlhdAndNB = (LogLikelihood,NegativeBinomial)
+type LlhdAndNB = (LogLikelihood, NegativeBinomial)
 
-type LlhdCalcState = (LlhdAndNB
-                     ,AbsoluteTime
-                     ,NumLineages)
+type LlhdCalcState = (LlhdAndNB, AbsoluteTime, NumLineages)